+++ TTKCore/musicNetworkKits/kg/musicdownloadquerykgplaylistthread.cpp
@@ -6,6 +6,7 @@
 MusicDownLoadQueryKGPlaylistThread::MusicDownLoadQueryKGPlaylistThread(QObject *parent)
     : MusicDownLoadQueryThreadAbstract(parent)
 {
+    m_pageSize = 30;
     m_queryServer = "Kugou";
 }
 
@@ -22,14 +23,22 @@ void MusicDownLoadQueryKGPlaylistThread::startSearchSong(QueryType type, const Q
     }
     else
     {
-        startSearchSongAll(playlist);
+        m_searchText = playlist;
+        startSearchSong(0);
     }
 }
 
-void MusicDownLoadQueryKGPlaylistThread::startSearchSongAll(const QString &type)
+void MusicDownLoadQueryKGPlaylistThread::startSearchSong(int offset)
 {
-    QUrl musicUrl = MusicCryptographicHash::decryptData(KG_PLAYLIST_URL, URL_KEY).arg(type);
+    if(!m_manager)
+    {
+        return;
+    }
+
     deleteAll();
+    m_pageTotal = 0;
+    QUrl musicUrl = MusicCryptographicHash::decryptData(KG_PLAYLIST_URL, URL_KEY)
+                    .arg(offset + 1).arg(m_pageSize).arg(m_searchText);
 
     QNetworkRequest request;
     request.setUrl(musicUrl);
@@ -47,6 +56,11 @@ void MusicDownLoadQueryKGPlaylistThread::startSearchSongAll(const QString &type)
 
 void MusicDownLoadQueryKGPlaylistThread::startSearchSong(const QString &playlist)
 {
+    if(!m_manager)
+    {
+        return;
+    }
+
     QUrl musicUrl = MusicCryptographicHash::decryptData(KG_PLAYLIST_ATTR_URL, URL_KEY).arg(playlist);
 
     QNetworkRequest request;
@@ -76,6 +90,12 @@ void MusicDownLoadQueryKGPlaylistThread::downLoadFinished()
     if(m_reply->error() == QNetworkReply::NoError)
     {
         QByteArray bytes = m_reply->readAll(); ///Get all the data obtained by request
+
+        QString buffer = QString(bytes);
+        buffer = buffer.split("global = ").back().split("total:").back();
+        buffer = buffer.split(",").front().remove("'").trimmed();
+        m_pageTotal = buffer.toInt();
+
         bytes = QString(bytes).split("global.special = ").back().split("];").front().toUtf8() + "]";
 
         QJson::Parser parser;
@@ -118,7 +138,7 @@ void MusicDownLoadQueryKGPlaylistThread::getDetailsFinished()
     emit clearAllItems();      ///Clear origin items
     m_musicSongInfos.clear();  ///Empty the last search to songsInfo
 
-    if(reply && reply->error() == QNetworkReply::NoError)
+    if(reply && m_manager &&reply->error() == QNetworkReply::NoError)
     {
         QByteArray bytes = reply->readAll();
