+++ TTKCore/musicNetworkKits/xm/musicdownloadqueryxmthread.cpp
@@ -8,7 +8,6 @@
 MusicDownLoadQueryXMThread::MusicDownLoadQueryXMThread(QObject *parent)
     : MusicDownLoadQueryThreadAbstract(parent)
 {
-    m_index = 0;
     m_queryServer = "XiaMi";
 }
 
@@ -19,16 +18,20 @@ QString MusicDownLoadQueryXMThread::getClassName()
 
 void MusicDownLoadQueryXMThread::startSearchSong(QueryType type, const QString &text)
 {
+    if(!m_manager)
+    {
+        return;
+    }
+
     m_searchText = text.trimmed();
     m_currentType = type;
-    QUrl musicUrl = MusicCryptographicHash::decryptData(XM_SONG_SEARCH_URL, URL_KEY).arg(text).arg(0).arg(50);
     deleteAll();
 
     QNetworkRequest request;
-    request.setUrl(musicUrl);
+    makeTokenQueryUrl(m_manager, &request,
+                      MusicCryptographicHash::decryptData(XM_SONG_DATA_URL, URL_KEY).arg(text).arg(1).arg(30),
+                      MusicCryptographicHash::decryptData(XM_SONG_URL, URL_KEY));
     request.setRawHeader("Content-Type", "application/x-www-form-urlencoded");
-    request.setRawHeader("Origin", MusicCryptographicHash::decryptData(XM_BASE_URL, URL_KEY).toUtf8());
-    request.setRawHeader("Referer", MusicCryptographicHash::decryptData(XM_BASE_URL, URL_KEY).toUtf8());
 #ifndef QT_NO_SSL
     QSslConfiguration sslConfig = request.sslConfiguration();
     sslConfig.setPeerVerifyMode(QSslSocket::VerifyNone);
@@ -41,7 +44,7 @@ void MusicDownLoadQueryXMThread::startSearchSong(QueryType type, const QString &
 
 void MusicDownLoadQueryXMThread::downLoadFinished()
 {
-    if(m_reply == nullptr)
+    if(!m_reply || !m_manager)
     {
         deleteAll();
         return;
@@ -49,12 +52,13 @@ void MusicDownLoadQueryXMThread::downLoadFinished()
 
     emit clearAllItems();      ///Clear origin items
     m_musicSongInfos.clear();  ///Empty the last search to songsInfo
-    m_songIds.clear();
-    m_index = 0;
 
     if(m_reply->error() == QNetworkReply::NoError)
     {
         QByteArray bytes = m_reply->readAll();///Get all the data obtained by request
+        bytes = bytes.replace("xiami(", "");
+        bytes = bytes.replace("callback(", "");
+        bytes.chop(1);
 
         QJson::Parser parser;
         bool ok;
@@ -62,8 +66,9 @@ void MusicDownLoadQueryXMThread::downLoadFinished()
         if(ok)
         {
             QVariantMap value = data.toMap();
-            if(value.contains("state") && value["state"].toInt() == 0 && value.contains("data"))
+            if(value.contains("data"))
             {
+                value = value["data"].toMap();
                 value = value["data"].toMap();
                 QVariantList datas = value["songs"].toList();
                 foreach(const QVariant &var, datas)
@@ -75,112 +80,40 @@ void MusicDownLoadQueryXMThread::downLoadFinished()
 
                     value = var.toMap();
                     MusicObject::MusicSongInfomation musicInfo;
-                    musicInfo.m_songId = QString::number(value["song_id"].toULongLong());
-                    musicInfo.m_artistId = QString::number(value["artist_id"].toULongLong());
-                    musicInfo.m_albumId = QString::number(value["album_id"].toULongLong());
-                    if(!m_querySimplify)
-                    {
-                        m_songIds << musicInfo.m_songId;
-                    }
-                    else
-                    {
-                        m_musicSongInfos << musicInfo;
-                    }
-                }
-            }
-        }
-
-        if(!m_querySimplify && !m_songIds.isEmpty())
-        {
-            m_musicSongInfos.clear();
-            startSongListQuery();
-        }
-        else
-        {
-            emit downLoadDataChanged(QString());
-            deleteAll();
-        }
-    }
-    else
-    {
-        emit downLoadDataChanged(QString());
-        deleteAll();
-    }
-}
-
-void MusicDownLoadQueryXMThread::songListFinished()
-{
-    QNetworkReply *reply = MObject_cast(QNetworkReply*, QObject::sender());
-    if(reply && reply->error() == QNetworkReply::NoError)
-    {
-        QByteArray bytes = reply->readAll();
-
-        QJson::Parser parser;
-        bool ok;
-        QVariant data = parser.parse(bytes, &ok);
-        if(ok)
-        {
-            QVariantMap value = data.toMap();
-            if(value.contains("data"))
-            {
-                ++m_index;
-                value = value["data"].toMap();
-                QVariantList datas = value["trackList"].toList();
-                foreach(const QVariant &var, datas)
-                {
-                    if(var.isNull())
-                    {
-                        continue;
-                    }
-
-                    value = var.toMap();
-                    MusicObject::MusicSongInfomation musicInfo;
                     musicInfo.m_singerName = value["singers"].toString();
                     musicInfo.m_songName = value["songName"].toString();
-                    musicInfo.m_timeLength = MusicTime::msecTime2LabelJustified(value["length"].toInt()*1000);
+                    musicInfo.m_timeLength = MusicTime::msecTime2LabelJustified(value["length"].toInt());
 
                     if(m_currentType != MovieQuery)
                     {
-                        musicInfo.m_songId = QString::number(value["songId"].toULongLong());
-                        musicInfo.m_artistId = QString::number(value["artistId"].toULongLong());
-                        musicInfo.m_smallPicUrl = MusicCryptographicHash::decryptData(XM_SONG_PIC_URL, URL_KEY) +
-                                                  value["album_logo"].toString().replace("_1.", "_4.");
-                        QString lrcUrl = value["lyric_url"].toString();
-                        if(!lrcUrl.endsWith("txt"))
+                        musicInfo.m_songId = value["songId"].toString();
+                        musicInfo.m_artistId = value["artistId"].toString();
+                        musicInfo.m_albumId = value["albumId"].toString();
+                        if(!m_querySimplify)
                         {
-                            musicInfo.m_lrcUrl = lrcUrl;
-                        }
+                            musicInfo.m_smallPicUrl = value["albumLogo"].toString();
 
-                        ///music normal songs urls
-                        QVariantList auditions = value["allAudios"].toList();
-                        foreach(const QVariant &audition, auditions)
-                        {
-                            QVariantMap audUrlsValue = audition.toMap();
-                            if(audUrlsValue.isEmpty())
+                            readFromMusicSongLrc(&musicInfo, m_manager, musicInfo.m_songId);
+                            readFromMusicSongAttribute(&musicInfo, value["listenFiles"],
+                                                       m_searchQuality, m_queryAllRecords);
+                            if(musicInfo.m_songAttrs.isEmpty())
                             {
                                 continue;
                             }
 
-                            readFromMusicSongAttribute(&musicInfo, audUrlsValue, m_searchQuality, m_queryAllRecords);
+                            MusicSearchedItem item;
+                            item.m_songname = musicInfo.m_songName;
+                            item.m_artistname = musicInfo.m_singerName;
+                            item.m_time = musicInfo.m_timeLength;
+                            item.m_type = mapQueryServerString();
+                            emit createSearchedItems(item);
                         }
-
-                        if(musicInfo.m_songAttrs.isEmpty())
-                        {
-                            continue;
-                        }
-
-                        MusicSearchedItem item;
-                        item.m_songname = musicInfo.m_songName;
-                        item.m_artistname = musicInfo.m_singerName;
-                        item.m_time = musicInfo.m_timeLength;
-                        item.m_type = mapQueryServerString();
-                        emit createSearchedItems(item);
                         m_musicSongInfos << musicInfo;
                     }
                     else
                     {
-                        //mv
-                        musicInfo.m_songId = value["mvUrl"].toString();
+                        //MV
+                        musicInfo.m_songId = value["mvId"].toString();
                         readFromMusicMVInfoAttribute(&musicInfo, musicInfo.m_songId, "mp4");
 
                         if(musicInfo.m_songAttrs.isEmpty())
@@ -197,53 +130,30 @@ void MusicDownLoadQueryXMThread::songListFinished()
                         m_musicSongInfos << musicInfo;
                     }
                 }
-
-                if(m_index >= m_songIds.count())
-                {
-                    foundOtherMovie();
-                    return;
-                }
-            }
-            else
-            {
-                foundOtherMovie();
             }
         }
-        else
-        {
-            foundOtherMovie();
-        }
     }
-    else
-    {
-        foundOtherMovie();
-    }
-}
 
-void MusicDownLoadQueryXMThread::startSongListQuery()
-{
-    foreach(const QString &id, m_songIds)
+    ///extra yyt movie
+    if(m_queryExtraMovie && m_currentType == MovieQuery)
     {
-        QNetworkRequest request;
-        request.setUrl(QUrl(MusicCryptographicHash::decryptData(XM_ARTIST_PLAYLIST_URL, URL_KEY).arg(id)));
-        request.setRawHeader("Content-Type", "application/x-www-form-urlencoded");
-        request.setRawHeader("Origin", MusicCryptographicHash::decryptData(XM_BASE_URL, URL_KEY).toUtf8());
-        request.setRawHeader("Referer", MusicCryptographicHash::decryptData(XM_BASE_URL, URL_KEY).toUtf8());
-    #ifndef QT_NO_SSL
-        QSslConfiguration sslConfig = request.sslConfiguration();
-        sslConfig.setPeerVerifyMode(QSslSocket::VerifyNone);
-        request.setSslConfiguration(sslConfig);
-    #endif
-        QNetworkReply *reply = m_manager->get(request);
-        connect(reply, SIGNAL(finished()), SLOT(songListFinished()) );
-        connect(reply, SIGNAL(error(QNetworkReply::NetworkError)), SLOT(replyError(QNetworkReply::NetworkError)) );
+        MusicSemaphoreLoop loop;
+        MusicDownLoadQueryYYTThread *yyt = new MusicDownLoadQueryYYTThread(this);
+        connect(yyt, SIGNAL(createSearchedItems(MusicSearchedItem)), SIGNAL(createSearchedItems(MusicSearchedItem)));
+        connect(yyt, SIGNAL(downLoadDataChanged(QString)), &loop, SLOT(quit()));
+        yyt->startSearchSong(MusicDownLoadQueryYYTThread::MovieQuery, m_searchText);
+        loop.exec();
+        m_musicSongInfos << yyt->getMusicSongInfos();
     }
+
+    emit downLoadDataChanged(QString());
+    deleteAll();
 }
 
 void MusicDownLoadQueryXMThread::readFromMusicMVInfoAttribute(MusicObject::MusicSongInfomation *info,
                                                               const QString &id, const QString &format)
 {
-    if(id.isEmpty())
+    if(id.isEmpty() || !m_manager)
     {
         return;
     }
@@ -258,8 +168,6 @@ void MusicDownLoadQueryXMThread::readFromMusicMVInfoAttribute(MusicObject::Music
     QNetworkRequest request;
     request.setUrl(musicUrl);
     request.setRawHeader("Content-Type", "application/x-www-form-urlencoded");
-    request.setRawHeader("Origin", MusicCryptographicHash::decryptData(XM_BASE_URL, URL_KEY).toUtf8());
-    request.setRawHeader("Referer", MusicCryptographicHash::decryptData(XM_BASE_URL, URL_KEY).toUtf8());
 #ifndef QT_NO_SSL
     QSslConfiguration sslConfig = request.sslConfiguration();
     sslConfig.setPeerVerifyMode(QSslSocket::VerifyNone);
@@ -283,22 +191,4 @@ void MusicDownLoadQueryXMThread::readFromMusicMVInfoAttribute(MusicObject::Music
         }
     }
     info->m_songAttrs.append(attr);
-}
-
-void MusicDownLoadQueryXMThread::foundOtherMovie()
-{
-    ///extra yyt movie
-    if(m_currentType == MovieQuery)
-    {
-        MusicSemaphoreLoop loop;
-        MusicDownLoadQueryYYTThread *yyt = new MusicDownLoadQueryYYTThread(this);
-connect(yyt, SIGNAL(createSearchedItems(MusicSearchedItem)), SIGNAL(createSearchedItems(MusicSearchedItem)));
-        connect(yyt, SIGNAL(downLoadDataChanged(QString)), &loop, SLOT(quit()));
-        yyt->startSearchSong(MusicDownLoadQueryYYTThread::MovieQuery, m_searchText);
-        loop.exec();
-        m_musicSongInfos << yyt->getMusicSongInfos();
-    }
-
-    emit downLoadDataChanged(QString());
-    deleteAll();
 }
